pyproject.toml
[tool.poetry]
name = "alex-blog"
version = "1.0.0"
description = "My personal blog"
authors = ["Alexander Dmitriev <alexander.dmitriev2580@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0"
django-pagedown = "^2.0"
django-markdownify = "^0.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
